+++ common/src/main/java/mca/entity/ai/brain/tasks/chore/AbstractChoreTask.java
@@ -41,7 +41,10 @@ public abstract class AbstractChoreTask extends Task<VillagerEntityMCA> {
 
             return false;
         }
-        return true;
+
+        long time = world.getTimeOfDay() % 24000;
+
+        return time > 2000 && time < 12000;
     }
 
     @Override